Tests/LibWeb/Text/expected/HTML/HTMLInputElement-type-state-change.txt

Changing type from "text" to "email" changes value from "  test  " to "test"
Changing type from "text" to "URL" changes value from "  test  " to "test"
Changing type from "hidden" to "text" changes value from "  t\re\nst  " to "  test  "
Changing type from "hidden" to "search" changes value from "  t\re\nst  " to "  test  "
Changing type from "hidden" to "tel" changes value from "  t\re\nst  " to "  test  "
Changing type from "hidden" to "password" changes value from "  t\re\nst  " to "  test  "
Changing type from "hidden" to "email" changes value from "  t\re\nst  " to "test"
Changing type from "hidden" to "URL" changes value from "   t\re\nst  " to "test"
Changing type from "text" to "number" changes value from "123.45" to "123.45"
Changing type from "text" to "number" changes value from "not-a-number" to ""
Changing type from "text" to "date" changes value from "not-a-date" to ""
Changing type from "text" to "date" changes value from "2024-09-10" to "2024-09-10"
Changing type from "text" to "date" changes value from "2024--09--10" to ""
Changing type from "text" to "datetime-local" changes value from "not-a-local-datetime" to ""
Changing type from "text" to "datetime-local" changes value from "2024-09-10 23:59" to "2024-09-10T23:59"
Changing type from "text" to "datetime-local" changes value from "2024-09-10T23:59" to "2024-09-10T23:59"
Changing type from "text" to "datetime-local" changes value from "2024--09--10TT23::59" to ""
Changing type from "text" to "month" changes value from "not-a-month" to ""
Changing type from "text" to "month" changes value from "2024-09" to "2024-09"
Changing type from "text" to "month" changes value from "2024--09" to ""
Changing type from "text" to "week" changes value from "not-a-week" to ""
Changing type from "text" to "week" changes value from "2024-W26" to "2024-W26"
Changing type from "text" to "week" changes value from "2024--W26" to ""
Changing type from "text" to "time" changes value from "not-a-time" to ""
Changing type from "text" to "time" changes value from "23:59:59" to "23:59:59"
Changing type from "text" to "time" changes value from "23::59::59" to ""
